requirements.txt
astropy>=3.2.1
backcall>=0.1.0
cycler>=0.10.0
decorator>=4.4.2
Fiona>=1.8.17
geopandas>=0.8.0
imageio>=2.5.0
ipython>=7.3.0
ipython-genutils>=0.2.0
joblib>=0.14.0
kiwisolver>=1.1.0
line-profiler>=2.1.2
matplotlib>=3.2.1
netCDF4>=1.5.4
numpy>=1.18.2
pandas>=1.0.0
pexpect>=4.7.0
pickleshare>=0.7.5
ply>=3.11
prompt-toolkit>=2.0.9
psutil>=5.6.3
ptyprocess>=0.6.0
Pygments>=2.4.2
pyparsing>=2.4.2
pyproj>=2.2.0
pyshp>=2.1.2
python-dateutil>=2.8.0
pytz>=2019.3
scikit-learn>=0.22.1
scipy>=1.4.1
seaborn>=0.11.0
Shapely>=1.6.4.post2
six>=1.12.0
spiceypy>=3.1.1
statsmodels>=0.10.1
tqdm>=4.46.0
traitlets>=4.3.3
wcwidth>=0.1.7
xarray>=0.14.0
Pillow>=8.0.1
pip>=20.3.1
attrs>=19.2.0
parso>=0.7.1
jedi>=0.17.2
click>=7.1.2
cligj>=0.7.1
munch>=2.5.0
setuptools>=41.2.0
patsy>=0.5.1
certifi>=2020.12.5
threadpoolctl>=2.1.0
submitit
wget
rioxarray
Rtree
pyarrow
